API reducer start
