+++ src/pages/ProfilePage.tsx
@@ -29,11 +29,11 @@ import { Edit } from '@material-ui/icons';
 
 import { Spacer } from '@pxblue/react-components';
 import { setUserData, setUserNotifications, UserDataChange } from '../redux/actions/user-data-actions';
-import { RESET_SET_USER_DATA } from '../redux/actions/types';
 
 import { ChangePassword } from '../components/dialogs/ChangePassword';
 import { Section } from '../components/display/Section';
 import { InfoCard } from '../components/display/InfoCard';
+import { RESET_API_STATUS } from '../redux/actions/types';
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
@@ -114,7 +114,7 @@ export const ProfilePage: React.FC = () => {
                                     backgroundColor: theme.palette.primary.light,
                                     backgroundImage: `url(https://www.swingessentials.com/images/profiles/${
                                         user.image ? `${user.username}/image.jpeg` : 'blank.png'
-                                        }`,
+                                    }`,
                                 }}
                             >
                                 <IconButton color={'inherit'}>
@@ -195,7 +195,7 @@ export const ProfileForm: React.FC = () => {
     const auth = useSelector((state: AppState) => state.auth);
     const user = useSelector((state: AppState) => state.user);
     const settings = useSelector((state: AppState) => state.settings);
-    const update = user.update;
+    const update = null; //useSelector((state: AppState) => state.status.);
 
     const [first, setFirst] = useState('');
     const [last, setLast] = useState('');
@@ -216,7 +216,7 @@ export const ProfileForm: React.FC = () => {
 
     useEffect(() => {
         if (changes) {
-            dispatch({ type: RESET_SET_USER_DATA });
+            dispatch({ type: RESET_API_STATUS.SET_USER_CHECKS });
             // dispatch({ type: RESET_SET_SETTINGS });
         }
     }, [changes, dispatch]);
@@ -238,7 +238,9 @@ export const ProfileForm: React.FC = () => {
             <Typography variant={'h5'}>Welcome to the Swing Essentials family!</Typography>
             <Typography paragraph>Help us get to know you by filling out your profile below.</Typography>
 
-            <Typography variant={'subtitle1'} style={{ fontWeight: 600, marginTop: 16, marginBottom: 16 }}>About Me:</Typography>
+            <Typography variant={'subtitle1'} style={{ fontWeight: 600, marginTop: 16, marginBottom: 16 }}>
+                About Me:
+            </Typography>
             <StyledTextField
                 label={'First Name'}
                 name={'first-name'}
@@ -279,49 +281,58 @@ export const ProfileForm: React.FC = () => {
                 disabled
                 label={`Email Address ${
                     auth.role !== 'customer' && auth.role !== 'administrator' ? '(unverified)' : ''
-                    }`}
+                }`}
                 name={'email'}
                 value={email}
                 onChange={(e): void => {
                     setEmail(e.target.value.substr(0, 128));
                 }}
             />
-            <Typography variant={'subtitle1'} style={{ fontWeight: 600, marginTop: 16, marginBottom: 16 }}>Notifications:</Typography>
-            <FormControlLabel labelPlacement="start"
-                control={<Switch checked={sendEmails} color={'default'} onChange={(e) => setSendEmails(e.target.checked)} />}
+            <Typography variant={'subtitle1'} style={{ fontWeight: 600, marginTop: 16, marginBottom: 16 }}>
+                Notifications:
+            </Typography>
+            <FormControlLabel
+                labelPlacement="start"
+                control={
+                    <Switch checked={sendEmails} color={'default'} onChange={(e) => setSendEmails(e.target.checked)} />
+                }
                 label="New Lesson Emails"
                 style={{ marginLeft: 0 }}
             />
 
             <div style={{ textAlign: 'center', marginTop: 16, minHeight: 36 }}>
-                {changes && update === 'unset' &&
+                {changes && update === 'unset' && (
                     <Button
                         color={'primary'}
                         variant={'contained'}
-                        onClick={changes ? (): void => {
-                            let newChanges: UserDataChange = {};
-                            if (first !== user.firstName) newChanges.firstName = first;
-                            if (last !== user.lastName) newChanges.lastName = last;
-                            if (location !== user.location) newChanges.location = location;
-                            if (phone !== user.phone) newChanges.phone = phone;
+                        onClick={
+                            changes
+                                ? (): void => {
+                                      let newChanges: UserDataChange = {};
+                                      if (first !== user.firstName) newChanges.firstName = first;
+                                      if (last !== user.lastName) newChanges.lastName = last;
+                                      if (location !== user.location) newChanges.location = location;
+                                      if (phone !== user.phone) newChanges.phone = phone;
 
-                            if (Object.keys(newChanges).length > 0) {
-                                console.log('changing personal');
-                                dispatch(setUserData(newChanges));
-                            }
-                            if (email !== user.email) {
-                                console.log('changing email');
-                                // dispatch change email request
-                            }
-                            if (sendEmails !== settings.notifications) {
-                                console.log('changing notification');
-                                dispatch(setUserNotifications({ subscribe: sendEmails }))
-                            }
-                        } : undefined}
+                                      if (Object.keys(newChanges).length > 0) {
+                                          console.log('changing personal');
+                                          dispatch(setUserData(newChanges));
+                                      }
+                                      if (email !== user.email) {
+                                          console.log('changing email');
+                                          // dispatch change email request
+                                      }
+                                      if (sendEmails !== settings.notifications) {
+                                          console.log('changing notification');
+                                          dispatch(setUserNotifications({ subscribe: sendEmails }));
+                                      }
+                                  }
+                                : undefined
+                        }
                     >
                         Save Changes
                     </Button>
-                }
+                )}
                 {update === 'pending' && <CircularProgress color={'inherit'} />}
                 {update === 'success' && <Typography variant={'caption'}>Success!</Typography>}
                 {update === 'error' && <Typography variant={'caption'}>Failed to update profile</Typography>}
